# backend/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "INVENTRAK Inventory Management API",
    "description": "Backend API for the INVENTRAK inventory management system. Both the SQLite (Express) backend and the npm-free fallback expose this identical contract. Live docs: GET /api/docs",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:4001",
      "description": "Local backend"
    }
  ],
...
  "paths": {
...
    "/api/products": {
      "get": {
        "tags": [
          "Products"
        ],
        "operationId": "listProducts",
        "summary": "List products (optionally paginated, searched, filtered)",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1
            },
            "description": "Page number (enables paginated response)"
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100
            },
            "description": "Items per page"
          },
          {
            "name": "search",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Search name/category/brand"
          },
          {
            "name": "category",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Filter by exact category"
          }
        ],
        "responses": {
          "200": {
            "description": "List of products (plain array, or { data, pagination } when page/limit provided)",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Product"
                      }
                    },
                    {
                      "$ref": "#/components/schemas/PaginatedProducts"
                    }
                  ]
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Products"
        ],
        "operationId": "createProduct",
        "summary": "Create a product (admin only)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Product created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreatedId"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
    "/api/products/bulk-prices": {
      "post": {
        "tags": [
          "Products"
        ],
        "operationId": "bulkUpdatePrices",
        "summary": "Set prices for many products in one request (admin only) \u00e2\u20ac\u201d the price-list CSV import",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkPriceInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Batch result: how many were updated and which entries were skipped (with reasons)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BulkPriceResult"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
    "/api/products/costs": {
      "get": {
        "tags": [
          "Products"
        ],
        "operationId": "listProductCosts",
        "summary": "Cost-of-goods sheet (admin only) \u2014 the only read path that reveals product cost",
        "description": "Cost of goods is the business's margin, so it is stripped from every public product read (catalog list, single product, QR lookup, tag page) regardless of who is asking. That catalog is served with `Cache-Control: public, max-age=300`, so an admin's response could otherwise sit in a shared cache and be read by an anonymous visitor. This endpoint is authenticated, admin-tier only, and never publicly cached.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "One row per product with its cost",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "required": [
                      "id",
                      "name",
                      "price"
                    ],
                    "properties": {
                      "id": {
                        "type": "integer"
                      },
                      "sku": {
                        "type": "string",
                        "nullable": true
                      },
                      "name": {
                        "type": "string"
                      },
                      "price": {
                        "type": "number"
                      },
                      "cost": {
                        "type": "number",
                        "nullable": true,
                        "description": "Cost of goods; null means the product is not costed."
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          }
        }
      }
    },
...
    "/api/products/{id}": {
      "get": {
        "tags": [
          "Products"
        ],
        "operationId": "getProduct",
        "summary": "Get a single product",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Product",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "put": {
        "tags": [
          "Products"
        ],
        "operationId": "updateProduct",
        "summary": "Update a product (admin only)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Update acknowledged",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Ok"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "delete": {
        "tags": [
          "Products"
        ],
        "operationId": "deleteProduct",
        "summary": "Soft-delete a product (admin only)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Product deactivated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
    "/api/products/qr/{code}": {
      "get": {
        "tags": [
          "Products"
        ],
        "operationId": "getProductByQr",
        "summary": "QR product lookup: resolve a scanned tag payload / SKU / bare id to a product (staff or admin)",
        "description": "The single round-trip behind QR product identification. Accepts the tag payload (INVENTRAK:PROD:<id>), a SKU (PRD-000042 or a custom code), or a bare numeric id (legacy barcode fallback). Returns the product, its QR tag data, a live per-location stock snapshot and open FIFO lots. The QR authorizes NOTHING \u00e2\u20ac\u201d normal authentication, product status, location, quantity and approval rules all still apply after the scan.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "maxLength": 120
            },
            "description": "Scanned QR content: tag payload, SKU or bare product id. URL-encoded when it contains colons."
          }
        ],
        "responses": {
          "200": {
            "description": "Product identified with live stock and lots",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QrProductLookup"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Validation failed or malformed input",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Missing or invalid token",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "Forbidden": {
        "description": "Admin access required",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
      "Conflict": {
        "description": "Conflict",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string"
          },
          "details": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "Product": {
        "type": "object",
        "required": [
          "id",
          "name",
          "category",
          "price",
          "status"
        ],
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "brand": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "size": {
            "type": "string"
          },
          "unit": {
            "type": "string"
          },
          "price": {
            "type": "number"
          },
          "cost": {
            "type": "number",
            "nullable": true,
            "minimum": 0,
            "description": "Cost of goods \u2014 what the caf\u00e9 pays, as opposed to price (what the customer pays). Null means the product is not costed; order costing then records cost_basis='none'/'imputed' rather than inventing a figure."
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "inactive"
            ]
          },
          "image": {
            "type": "string",
            "nullable": true,
            "description": "Product photo: '/images/<file>' served by the backend, or a full URL"
          },
          "created_at": {
            "type": "string"
          },
          "updated_at": {
            "type": "string"
          },
          "sku": {
            "type": "string",
            "nullable": true,
            "description": "Stable product identifier encoded in the printed QR tag (system default PRD-000001; custom codes allowed). The QR payload itself is INVENTRAK:PROD:<id> \u00e2\u20ac\u201d the sku is display/back-compat only."
          }
        }
      },
      "ProductInput": {
        "type": "object",
        "required": [
          "name",
          "category",
          "price"
        ],
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 200
          },
          "category": {
            "type": "string",
            "maxLength": 100
          },
          "brand": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "size": {
            "type": "string"
          },
          "unit": {
            "type": "string"
          },
          "price": {
            "type": "number",
            "minimum": 0
          },
          "cost": {
            "type": "number",
            "nullable": true,
            "minimum": 0,
            "description": "Cost of goods (optional, nullable). Omit or null to mark the product as not costed."
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "inactive"
            ]
          },
          "image": {
            "type": "string",
            "maxLength": 300,
            "description": "Product photo: '/images/<file>' served by the backend, or a full URL"
          },
          "sku": {
            "type": "string",
            "description": "Custom SKU (3-32 chars: A-Z, 0-9, dashes). Omit to get the deterministic system SKU (PRD-<padded id>). Cannot be cleared once set \u00e2\u20ac\u201d reprint the tag instead."
          }
        }
      },
...
      "BulkPriceEntry": {
        "type": "object",
        "required": [
          "name",
          "price"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "minimum": 1,
            "description": "Optional product id; matched first when present"
          },
          "name": {
            "type": "string",
            "maxLength": 200,
            "description": "Product name; matched case-insensitively when no id matches"
          },
          "price": {
            "type": "number",
            "minimum": 0
          }
        }
      },
      "BulkPriceInput": {
        "type": "object",
        "required": [
          "prices"
        ],
        "properties": {
          "prices": {
            "type": "array",
            "minItems": 1,
            "maxItems": 2000,
            "items": {
              "$ref": "#/components/schemas/BulkPriceEntry"
            }
          }
        }
      },
      "BulkSkippedEntry": {
        "type": "object",
        "required": [
          "name",
          "reason"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "reason": {
            "type": "string",
            "enum": [
              "invalid price",
              "name too long",
              "not found"
            ]
          }
        }
      },
      "BulkPriceResult": {
        "type": "object",
        "required": [
          "ok",
          "total",
          "updated",
          "skipped"
        ],
        "properties": {
          "ok": {
            "type": "boolean"
          },
          "total": {
            "type": "integer",
            "minimum": 0
          },
          "updated": {
            "type": "integer",
            "minimum": 0
          },
          "skipped": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BulkSkippedEntry"
            }
          }
        }
      },
      "CreatedId": {
        "type": "object",
        "required": [
          "id"
        ],
        "properties": {
          "id": {
            "type": "integer"
          }
        }
      },
      "Ok": {
        "type": "object",
        "required": [
          "ok"
        ],
        "properties": {
          "ok": {
            "type": "boolean"
          }
        }
      },
      "Message": {
        "type": "object",
        "required": [
          "ok",
          "message"
        ],
        "properties": {
          "ok": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "notify": {
            "type": "object",
            "description": "Delivery status of the verification code (present only when a code was actually generated: register + resend-verification). email is always attempted; sms only when a phone was provided.",
            "properties": {
              "email": {
                "type": "boolean"
              },
              "sms": {
                "type": "boolean"
              }
            }
          }
        }
      },
...
      "Pagination": {
        "type": "object",
        "required": [
          "page",
          "limit",
          "total",
          "totalPages"
        ],
        "properties": {
          "page": {
            "type": "integer"
          },
          "limit": {
            "type": "integer"
          },
          "total": {
            "type": "integer"
          },
          "totalPages": {
            "type": "integer"
          }
        }
      },
      "PaginatedProducts": {
        "type": "object",
        "required": [
          "data",
          "pagination"
        ],
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Product"
            }
          },
          "pagination": {
            "$ref": "#/components/schemas/Pagination"
          }
        }
      },
...
      "QrProductLookup": {
        "type": "object",
        "required": [
          "product",
          "qr",
          "stock",
          "lots"
        ],
        "properties": {
          "product": {
            "$ref": "#/components/schemas/Product"
          },
          "qr": {
            "type": "object",
            "required": [
              "payload",
              "sku"
            ],
            "properties": {
              "payload": {
                "type": "string",
                "description": "INVENTRAK:PROD:<id> \u00e2\u20ac\u201d the exact string encoded in the printed tag; handy for reprinting."
              },
              "sku": {
                "type": "string",
                "description": "SKU shown on the label."
              }
            }
          },
          "stock": {
            "type": "object",
            "required": [
              "locations",
              "total",
              "status"
            ],
            "properties": {
              "locations": {
                "type": "object",
                "additionalProperties": {
                  "type": "number"
                },
                "description": "Quantity per storage-area name (Showroom, Stockroom 1, Stockroom 2)."
              },
              "total": {
                "type": "number"
              },
              "status": {
                "type": "string",
                "enum": [
                  "ok",
                  "low",
                  "out"
                ],
                "description": "Same bands as the Inventory page (total < 80 = low)."
              }
            }
          },
          "lots": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/StockLot"
            }
          }
        }
      }
    }
  }
}